requirements.txt
openai>=1.12.0
mcp>=0.9.0
customtkinter>=5.2.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pywin32>=306; sys_platform == "win32"
watchdog>=3.0.0
win10toast>=0.9
pyautogui>=0.9.54
pyperclip>=1.8.2
